in case of error use txn to make sure we do not miss data
