+++ queue/mysql/query.go
@@ -76,68 +76,3 @@ func (q *queueImpl) internalJobDetails(ctx context.Context, req queue.JobDetails
 
 	return
 }
-
-func (q *queueImpl) MarkJobCompletedWithRetry(ctx context.Context, req queue.MarkJobFailedWithRetryRequest) (result *queue.MarkJobFailedWithRetryResponse, err error) {
-	result = &queue.MarkJobFailedWithRetryResponse{Done: false}
-
-	// Get the partition time
-	part := time.Time{}
-	if part, err = q.idToTime(req.Id); err != nil {
-		return nil, errors.Wrap(err, "not able to get time out of id: id=%s", req.Id)
-	}
-	part = partitionBasedOnProcessAtTime(part)
-
-	var jobFetchResponse *queue.JobDetailsResponse
-	if jobFetchResponse, err = q.FetchJobDetails(ctx, queue.JobDetailsRequest{Id: req.Id}); err != nil {
-		return nil, errors.Wrap(err, "failed to get job with id=%s - needed to setup retry", req.Id)
-	}
-
-	if jobFetchResponse.RemainingExecution <= 0 {
-		if _, err = q.updateJobStatusStatement.ExecContext(ctx, queue.StatusFailed, queue.SubStatusNoRetryPendingError, req.Id, part); err != nil {
-			return nil, errors.Wrap(err, "failed to update the job to mark failed: id=%s", req.Id)
-		}
-		result.Done = true
-	} else {
-		var scheduleResponse *queue.ScheduleResponse
-		if scheduleResponse, err = q.Schedule(ctx, queue.ScheduleRequest{
-			At:                 req.ScheduleRetryAt,
-			JobType:            jobFetchResponse.JobType,
-			Tenant:             jobFetchResponse.Tenant,
-			CorrelationId:      jobFetchResponse.CorrelationId,
-			RemainingExecution: jobFetchResponse.RemainingExecution - 1,
-			StringUdf1:         jobFetchResponse.StringUdf1,
-			StringUdf2:         jobFetchResponse.StringUdf2,
-			IntUdf1:            jobFetchResponse.IntUdf1,
-			IntUdf2:            jobFetchResponse.IntUdf1,
-			Properties:         jobFetchResponse.Properties,
-		}); err != nil {
-			return nil, errors.Wrap(err, "failed to add new retry jobs (some retries are remaining for this job): id=%s", req.Id)
-		}
-
-		if _, err = q.updateJobStatusStatement.ExecContext(ctx, queue.StatusFailed, queue.SubStatusRetryPendingError, req.Id, part); err != nil {
-			return nil, errors.Wrap(err, "failed to update the job to mark failed: id=%s", req.Id)
-		}
-
-		result.RetryJobId = scheduleResponse.Id
-		result.Done = true
-	}
-
-	return
-}
-
-func (q *queueImpl) MarkJobCompleted(ctx context.Context, req queue.MarkJobCompletedRequest) (result *queue.MarkJobCompletedResponse, err error) {
-	result = &queue.MarkJobCompletedResponse{}
-
-	// Get the partition time
-	part := time.Time{}
-	if part, err = q.idToTime(req.Id); err != nil {
-		return nil, errors.Wrap(err, "not able to get time out of id: id=%s", req.Id)
-	}
-	part = partitionBasedOnProcessAtTime(part)
-
-	// Mark it done
-	if _, err = q.updateJobStatusStatement.ExecContext(ctx, queue.StatusDone, queue.SubStatusDone, req.Id, part); err != nil {
-		return nil, errors.Wrap(err, "failed to update the job: id=%s", req.Id)
-	}
-	return
-}
